+++ dropwizard-v5/src/test/scala/fr/davit/akka/http/metrics/dropwizard/DropwizardRegistrySpec.scala
@@ -17,13 +17,14 @@
 package fr.davit.akka.http.metrics.dropwizard
 
 import akka.http.scaladsl.model.StatusCodes
-import com.codahale.metrics.MetricRegistry
 import fr.davit.akka.http.metrics.core.Dimension
 import fr.davit.akka.http.metrics.core.HttpMetricsRegistry.{PathDimension, StatusGroupDimension}
+import io.dropwizard.metrics5.MetricName
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
 import scala.concurrent.duration._
+
 import scala.jdk.CollectionConverters._
 
 class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
@@ -33,21 +34,20 @@ class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
   trait Fixture {
     val registry = DropwizardRegistry()
 
-    def underlyingCounter(name: String, dim: Option[Dimension] = None): Long = {
-      registry.underlying.getCounters.asScala(metricName(name, dim)).getCount
+    def underlyingCounter(name: String, dims: Seq[Dimension] = Seq.empty): Long = {
+      registry.underlying.getCounters.asScala(metricName(name, dims)).getCount
     }
 
-    def underlyingHistogram(name: String, dim: Option[Dimension] = None): Long = {
-      registry.underlying.getHistograms.asScala(metricName(name, dim)).getSnapshot.getValues.sum
+    def underlyingHistogram(name: String, dims: Seq[Dimension] = Seq.empty): Long = {
+      registry.underlying.getHistograms.asScala(metricName(name, dims)).getSum
     }
 
-    def underlyingTimer(name: String, dim: Option[Dimension] = None): Long = {
-      registry.underlying.getTimers.asScala(metricName(name, dim)).getSnapshot.getValues.sum
+    def underlyingTimer(name: String, dims: Seq[Dimension] = Seq.empty): Long = {
+      registry.underlying.getTimers.asScala(metricName(name, dims)).getSum
     }
 
-    private def metricName(name: String, dim: Option[Dimension]): String = dim match {
-      case Some(d) => MetricRegistry.name(name, s"${d.value}-${d.key}")
-      case None    => MetricRegistry.name(name)
+    private def metricName(name: String, dims: Seq[Dimension]): MetricName = {
+      MetricName.build(name).tagged(dims.map(d => d.key -> d.value).toMap.asJava)
     }
   }
 
@@ -71,9 +71,7 @@ class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
     underlyingCounter("akka.http.responses") shouldBe 1L
 
     registry.responses.inc(dimensions)
-    dimensions.foreach { d =>
-      underlyingCounter("akka.http.responses", Some(d)) shouldBe 1L
-    }
+    underlyingCounter("akka.http.responses", dimensions) shouldBe 1L
   }
 
   it should "set responsesErrors metrics in the underlying registry" in new Fixture {
@@ -81,9 +79,7 @@ class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
     underlyingCounter("akka.http.responses.errors") shouldBe 1L
 
     registry.responsesErrors.inc(dimensions)
-    dimensions.foreach { d =>
-      underlyingCounter("akka.http.responses.errors", Some(d)) shouldBe 1L
-    }
+    underlyingCounter("akka.http.responses.errors", dimensions) shouldBe 1L
   }
 
   it should "set responsesDuration metrics in the underlying registry" in new Fixture {
@@ -91,9 +87,7 @@ class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
     underlyingTimer("akka.http.responses.duration") shouldBe 3000000000L
 
     registry.responsesDuration.observe(3.seconds, dimensions)
-    dimensions.foreach { d =>
-      underlyingTimer("akka.http.responses.duration", Some(d)) shouldBe 3000000000L
-    }
+    underlyingTimer("akka.http.responses.duration", dimensions) shouldBe 3000000000L
   }
 
   it should "set responsesSize metrics in the underlying registry" in new Fixture {
@@ -101,9 +95,7 @@ class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
     underlyingHistogram("akka.http.responses.bytes") shouldBe 3L
 
     registry.responsesSize.update(3, dimensions)
-    dimensions.foreach { d =>
-      underlyingHistogram("akka.http.responses.bytes", Some(d)) shouldBe 3L
-    }
+    underlyingHistogram("akka.http.responses.bytes", dimensions) shouldBe 3L
   }
 
   it should "set connectionsActive metrics in the underlying registry" in new Fixture {