+++ project/Dependencies.scala
@@ -6,21 +6,26 @@ object Dependencies {
     val akka                  = "2.6.10"
     val akkaHttp              = "10.2.2"
     val datadog               = "2.10.5"
-    val dropwizard            = "5.0.0"
+    val dropwizard            = "4.1.2"
+    val dropwizardV5          = "5.0.0"
     val enumeratum            = "1.6.1"
     val logback               = "1.2.3"
     val prometheus            = "0.9.0"
     val scalaCollectionCompat = "2.3.1"
+    val scalaLogging          = "3.9.2"
     val scalaMock             = "5.1.0"
     val scalaTest             = "3.2.3"
   }
 
-  val akkaHttp         = "com.typesafe.akka"      %% "akka-http"            % Versions.akkaHttp
-  val datadog          = "com.datadoghq"          % "java-dogstatsd-client" % Versions.datadog
-  val dropwizardCore   = "io.dropwizard.metrics5" % "metrics-core"          % Versions.dropwizard
-  val dropwizardJson   = "io.dropwizard.metrics5" % "metrics-json"          % Versions.dropwizard
-  val enumeratum       = "com.beachape"           %% "enumeratum"           % Versions.enumeratum
-  val prometheusCommon = "io.prometheus"          % "simpleclient_common"   % Versions.prometheus
+  val akkaHttp         = "com.typesafe.akka"          %% "akka-http"            % Versions.akkaHttp
+  val datadog          = "com.datadoghq"              % "java-dogstatsd-client" % Versions.datadog
+  val dropwizardCore   = "io.dropwizard.metrics"      % "metrics-core"          % Versions.dropwizard
+  val dropwizardJson   = "io.dropwizard.metrics"      % "metrics-json"          % Versions.dropwizard
+  val dropwizardV5Core = "io.dropwizard.metrics5"     % "metrics-core"          % Versions.dropwizardV5
+  val dropwizardV5Json = "io.dropwizard.metrics5"     % "metrics-json"          % Versions.dropwizardV5
+  val enumeratum       = "com.beachape"               %% "enumeratum"           % Versions.enumeratum
+  val prometheusCommon = "io.prometheus"              % "simpleclient_common"   % Versions.prometheus
+  val scalaLogging     = "com.typesafe.scala-logging" %% "scala-logging"        % Versions.scalaLogging
 
   object Provided {
     val akkaStream = "com.typesafe.akka" %% "akka-stream" % Versions.akka % "provided"
@@ -32,7 +37,8 @@ object Dependencies {
     val akkaSlf4j             = "com.typesafe.akka"      %% "akka-slf4j"              % Versions.akka                  % "it,test"
     val akkaStreamTestkit     = "com.typesafe.akka"      %% "akka-stream-testkit"     % Versions.akka                  % "it,test"
     val akkaTestkit           = "com.typesafe.akka"      %% "akka-testkit"            % Versions.akka                  % "it,test"
-    val dropwizardJvm         = "io.dropwizard.metrics5" % "metrics-jvm"              % Versions.dropwizard            % "it,test"
+    val dropwizardJvm         = "io.dropwizard.metrics"  % "metrics-jvm"              % Versions.dropwizard            % "it,test"
+    val dropwizardV5Jvm       = "io.dropwizard.metrics5" % "metrics-jvm"              % Versions.dropwizardV5          % "it,test"
     val logback               = "ch.qos.logback"         % "logback-classic"          % Versions.logback               % "it,test"
     val prometheusHotspot     = "io.prometheus"          % "simpleclient_hotspot"     % Versions.prometheus            % "it,test"
     val scalaCollectionCompat = "org.scala-lang.modules" %% "scala-collection-compat" % Versions.scalaCollectionCompat % "it,test"