+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+## v1.5.0 (2021-01-12)
+
+- Update dependencies
 - Split dropwizard v4 and v5. v4 being the default.
 - Fix bug when HttpEntity.Default is used with HTTP/1.0 protocol